Create component for welcome section on home page
